requirements.txt
requests
beautifulsoup4
lxml
pyyaml
python-dotenv
